# src/Spotless.Web/lib/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Spotless API",
    "description": "Laundry and dry cleaning service API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Payments/initiate": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Initiates a payment for an order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Payment.InitiatePaymentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Payment.InitiatePaymentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Payment.InitiatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Payment.InitiatePaymentResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Payment.InitiatePaymentResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Payment.InitiatePaymentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/Payments/webhook": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Processes Paymob payment webhook notifications",
        "parameters": [
          {
            "name": "Hmac-SHA512",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Interfaces.PaymobProcessedCallbackData"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Interfaces.PaymobProcessedCallbackData"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Interfaces.PaymobProcessedCallbackData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "Spotless.Application.Dtos.Payment.InitiatePaymentDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "format": "uuid"
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/Spotless.Domain.Enums.PaymentMethod"
          },
          "returnUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Spotless.Application.Dtos.Payment.InitiatePaymentResponseDto": {
        "type": "object",
        "properties": {
          "paymentId": {
            "type": "string",
            "format": "uuid"
          },
          "paymentUrl": {
            "type": "string",
            "nullable": true
          },
          "transactionReference": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Spotless.Application.Interfaces.PaymobProcessedCallbackData": {
        "type": "object",
        "properties": {
          "amountCents": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "errorOccured": {
            "type": "boolean"
          },
          "hasParentTransaction": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "integrationId": {
            "type": "integer",
            "format": "int32"
          },
          "is3dSecure": {
            "type": "boolean"
          },
          "isAuth": {
            "type": "boolean"
          },
          "isCapture": {
            "type": "boolean"
          },
          "isRefunded": {
            "type": "boolean"
          },
          "isStandalonePayment": {
            "type": "boolean"
          },
          "isVoided": {
            "type": "boolean"
          },
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "owner": {
            "type": "integer",
            "format": "int32"
          },
          "pending": {
            "type": "boolean"
          },
          "sourceDataPan": {
            "type": "string",
            "nullable": true
          },
          "sourceDataSubType": {
            "type": "string",
            "nullable": true
          },
          "sourceDataType": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Spotless.Domain.Enums.PaymentMethod": {
        "enum": [
          "CreditCard",
          "DebitCard",
          "Fawry",
          "PayPal",
          "Wallet",
          "CashOnDelivery",
          "Paymob",
          "Instapay"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}